// src/types/Cart/CartResponse.ts
// Generated by https://quicktype.io

export interface ICartResponse {
  isSuccessful: boolean;
  message: string;
  data: Data;
  errors: Error[];
}

export interface Error {
  message: string;
  code: string;
  productInfo: ProductInfoError;
}

export interface Data {
  id: null;
  user: User;
  status: string;
  shoppingCartAbandoned: boolean;
  updateCartBitacoraBD: boolean;
  cartUniqueId: string;
  createdAt: string;
  updatedAt: string;
  cartAttributes: CartAttributes;
}

export interface ProductInfoError {
  productId: string;
  productName: string;
  images: string[];
}

export interface CartAttributes {
  notes: null;
  total: string;
  totalItemCart: number;
  totalItems: number;
  delivery: number;
  discount: number;
  subTotal: number;
  fee: number;
  products: Product[];
}

export interface Product {
  id: string;
  quantity: number;
  productTotal: number;
  quantityAvailable: string;
  iva: number;
  stockDecreaseUpdate: boolean;
  attributes: Attributes;
}

export interface Attributes {
  sku: string;
  slug: string;
  images: string[];
  stock: number;
  value: number;
  quantityByCart: number;
  name: string;
  type: string;
  sellerName: string;
  sellerId: number;
  assumedValue: number;
  regularPrice: number;
  length: number;
  width: number;
  height: number;
  weight: number;
  warehouseId: number;
  checkTransport: boolean;
  categories: Category[];
  slugParent: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
}

export interface User {
  uniqueId: string;
  email: string;
  userName: string;
  phone: string;
  address: string;
}

// ***************************************************
//  Body for the POST request

export interface IBodyProduct {
  cart?: {
    cartUniqueId?: string;
    user_info?: {
      uniqueId?: string;
      email?: string;
      userName?: string;
      phone?: string;
      address?: string;
    };
    lastUserStep?: string;
    notes: string;
  };
  product: {
    productId: string;
    quantity: number;
  };
}

export interface IMetadataUpdateRequest {
  cartUniqueId?: string;
  delivery?: number;
  user_info?: Partial<User>;
  Fee?: number;
  lastUserStep?: string;
}